+++ AppCenter/AppCenter/Internals/Channel/MSACChannelUnitDefault.m
@@ -207,7 +207,7 @@ static NSString *const kMSACStartTimestampPrefix = @"ChannelStartTimer";
                 MSACLogWarning([MSACAppCenter logTag], @"Batch Id %@ not expected, ignore.", ingestionBatchId);
                 return;
               }
-              BOOL succeeded = response.statusCode == MSACHTTPCodesNo200OK;
+              BOOL succeeded = [MSACHttpUtil isSuccessStatusCode:response.statusCode];
               if (succeeded) {
                 MSACLogDebug([MSACAppCenter logTag], @"Log(s) sent with success, batch Id:%@.", ingestionBatchId);
 
@@ -566,7 +566,7 @@ static NSString *const kMSACStartTimestampPrefix = @"ChannelStartTimer";
   }
 }
 
-- (void)notifyFailureBeforeSendingForItem:(id<MSACLog>)item withError:(NSError *)error {
+- (void)notifyFailureBeforeSendingForItem:(id<MSACLog>)item withError:(nullable NSError *)error {
   NSArray *synchronizedDelegates;
   @synchronized(self.delegates) {