+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Version 4.1.2 (Under development)
 
+### App Center Crashes
+
+* **[Fix]** Fix error nullability in crashes delegate.
+
 ### App Center Distribute
 
 * **[Fix]** Fix linking framework `AuthenticationServices`.
+
 ___
 
 ## Version 4.1.1